To get unit test working for blank suite, restored line from before Diego Virasoro patch.'
git-svn-id: svn+ssh://rubyforge.org/var/svn/nasarb@425 15d20c99-a110-0410-8ff5-f949c1c7ce36

+++ funit/trunk/lib/funit/testsuite.rb
@@ -19,6 +19,7 @@ module Funit
       @line_number = 'blank'
       @suite_name = suite_name
       @suite_content = suite_content
+      return nil unless funit_exists?(suite_name)
       File.delete(suite_name+"_fun.f90") if File.exists?(suite_name+"_fun.f90")
       super(suite_name+"_fun.f90","w")
       @tests, @setup, @teardown = [], [], []
